app.py
#!/usr/bin/env python3
"""
Charlotte Land Map Backend
A Flask-based API for land parcel analysis and development opportunities
"""

import os
from flask import Flask, request, jsonify
from flask_cors import CORS
from dotenv import load_dotenv
import logging

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Initialize Flask app
app = Flask(__name__)
CORS(app)  # Enable CORS for frontend integration

# Configuration
app.config['SECRET_KEY'] = os.getenv('SECRET_KEY', 'dev-secret-key')
app.config['DATABASE_URL'] = os.getenv('DATABASE_URL', 'postgresql://localhost/charlotte_land_map')

@app.route('/')
def home():
    """Home endpoint"""
    return jsonify({
        'message': 'Charlotte Land Map Backend API',
        'version': '1.0.0',
        'status': 'running'
    })

@app.route('/api/health')
def health_check():
    """Health check endpoint"""
    return jsonify({'status': 'healthy', 'service': 'charlotte-land-map'})

@app.route('/api/parcels/search', methods=['POST'])
def search_parcels():
    """Search parcels based on criteria"""
    try:
        data = request.get_json()
        
        # Extract search parameters
        min_acres = data.get('min_acres', 0)
        max_acres = data.get('max_acres', None)
        zoning_type = data.get('zoning_type', None)
        vacant_only = data.get('vacant_only', False)
        road_frontage = data.get('road_frontage', None)
        slope_range = data.get('slope_range', None)
        
        # TODO: Implement actual database query logic
        # For now, return mock data
        mock_results = {
            'parcels': [
                {
                    'id': 'P001',
                    'address': '123 Main St, Charlotte, NC',
                    'acres': 5.2,
                    'zoning': 'R-3',
                    'vacant': True,
                    'road_frontage': 200,
                    'slope': '2-5%'
                }
            ],
            'total_count': 1,
            'search_criteria': data
        }
        
        return jsonify(mock_results)
        
    except Exception as e:
        logger.error(f"Error searching parcels: {str(e)}")
        return jsonify({'error': 'Internal server error'}), 500

@app.route('/api/parcels/<parcel_id>/details')
def get_parcel_details(parcel_id):
    """Get detailed information about a specific parcel"""
    try:
        # TODO: Implement actual database query logic
        mock_details = {
            'id': parcel_id,
            'address': '123 Main St, Charlotte, NC',
            'owner': {
                'name': 'John Doe',
                'phone': '555-123-4567',
                'email': 'john.doe@example.com'
            },
            'property_details': {
                'acres': 5.2,
                'zoning': 'R-3',
                'land_use': 'Residential',
                'road_frontage': 200,
                'slope': '2-5%',
                'utilities': ['Water', 'Sewer', 'Electric']
            },
            'development_potential': {
                'max_units': 8,
                'setback_requirements': '25ft front, 10ft sides, 25ft rear',
                'height_limit': '35ft'
            }
        }
        
        return jsonify(mock_details)
        
    except Exception as e:
        logger.error(f"Error getting parcel details: {str(e)}")
        return jsonify({'error': 'Internal server error'}), 500

@app.route('/api/owner/search', methods=['POST'])
def search_owner():
    """Search for owner contact information"""
    try:
        data = request.get_json()
        address = data.get('address')
        name = data.get('name')
        
        if not address and not name:
            return jsonify({'error': 'Address or name required'}), 400
        
        # TODO: Implement actual WhitePages/TruePeopleSearch API calls
        # For now, return mock data
        mock_owner_info = {
            'name': 'John Doe',
            'phone': '555-123-4567',
            'email': 'john.doe@example.com',
            'address': address or '123 Main St, Charlotte, NC',
            'source': 'Mock Data'
        }
        
        return jsonify(mock_owner_info)
        
    except Exception as e:
        logger.error(f"Error searching owner: {str(e)}")
        return jsonify({'error': 'Internal server error'}), 500

@app.route('/api/zoning/regulations')
def get_zoning_regulations():
    """Get zoning regulations and subdivision requirements"""
    try:
        # TODO: Implement actual zoning data retrieval
        mock_regulations = {
            'residential': {
                'R-1': {
                    'min_lot_size': '1 acre',
                    'max_density': '1 unit per acre',
                    'setbacks': '25ft front, 10ft sides, 25ft rear'
                },
                'R-3': {
                    'min_lot_size': '0.5 acres',
                    'max_density': '2 units per acre',
                    'setbacks': '20ft front, 8ft sides, 20ft rear'
                }
            },
            'commercial': {
                'C-1': {
                    'min_lot_size': '0.25 acres',
                    'max_coverage': '60%',
                    'setbacks': '15ft front, 5ft sides, 15ft rear'
                }
            }
        }
        
        return jsonify(mock_regulations)
        
    except Exception as e:
        logger.error(f"Error getting zoning regulations: {str(e)}")
        return jsonify({'error': 'Internal server error'}), 500

if __name__ == '__main__':
    port = int(os.getenv('PORT', 5000))
    debug = os.getenv('FLASK_ENV') == 'development'
    
    logger.info(f"Starting Charlotte Land Map Backend on port {port}")
    app.run(host='0.0.0.0', port=port, debug=debug)

